requirements.txt
spacy
transformers
requests
torch>=1.12.0
openai>=1.0.0
python-dotenv